+++ src/main/java/Batalla/Datos/Ninja.java
@@ -5,12 +5,12 @@ public class Ninja extends Personaje
     public Ninja()
     {
         super("Ninja",
-                100,
-                15,
-                15,
-                2,
-                0.30,
-                0.15,
-                0.10);
+                (float)100,
+                (float)15,
+                (float)15,
+                (float)2,
+                (float)0.30,
+                (float)0.15,
+                (float)0.10);
     }
 }
